Program now remembers listings it's searched when previously run
The program can now remember listings it's already found even in previous runs of the program.
Before, it would not show already found listings if you searched again in the same run of the program, but now it remembers all listings it's previously found in all runs as it records the URL of each listing to a txt file which it reads every time it's opened to create a dictionary of all URL's it has seen in order to avoid them when searching again.