Added folders, no file made when 0 results found
Added folders for the program files and results so that the folder for the program is more clean and results can be found easily.

Made it so that no new file is created when 0 results are found as that just makes an empty txt file which is not of much use, program tells user no results found and that not file was created.